move all interfaces and classes inside AbstractMap file to individual files and their respective folders categorize Abstract map with private fields, properties, methods in regions
