+++ sdkproject/Assets/Mapbox/Unity/DataContainers/MapScalingOptions.cs
@@ -1,4 +1,6 @@
-﻿namespace Mapbox.Unity.Map
+﻿using Mapbox.Unity.Map.Interfaces;
+
+namespace Mapbox.Unity.Map
 {
 	using System;
 	using UnityEngine;
