+++ sdkproject/Assets/Mapbox/Unity/Map/AbstractMapVisualizer.cs
@@ -1,3 +1,5 @@
+using Mapbox.Unity.Map.Interfaces;
+
 namespace Mapbox.Unity.Map
 {
 	using System.Linq;
