+++ sdkproject/Assets/Mapbox/Unity/Map/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d3f7157c1c1f94f4aaac6081e3449f9b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
