+++ sdkproject/Assets/Mapbox/Unity/Map/Interfaces/IMapPlacementStrategy.cs
@@ -1,0 +1,7 @@
+namespace Mapbox.Unity.Map.Interfaces
+{
+	public interface IMapPlacementStrategy
+	{
+		void SetUpPlacement(AbstractMap map);
+	}
+}
