+++ sdkproject/Assets/Mapbox/Unity/Map/Interfaces/IMapPlacementStrategy.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: aa6c9ee2f3b546dc9df96694c85e82c5
+timeCreated: 1538658253
