+++ sdkproject/Assets/Mapbox/Unity/Map/Interfaces/IMapScalingStrategy.cs
@@ -1,0 +1,7 @@
+namespace Mapbox.Unity.Map.Interfaces
+{
+	public interface IMapScalingStrategy
+	{
+		void SetUpScaling(AbstractMap map);
+	}
+}
