+++ sdkproject/Assets/Mapbox/Unity/Map/Interfaces/IMapScalingStrategy.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a3d0db333c6d474aa152ccac7439e444
+timeCreated: 1538658226
