+++ sdkproject/Assets/Mapbox/Unity/Map/Strategies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 817a1846479397d458272568e8bb804f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
