+++ sdkproject/Assets/Mapbox/Unity/Map/Strategies/MapPlacementAtLocationCenterStrategy.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d44d0606bd83495bab6d3da0a48ad090
+timeCreated: 1538658271
