+++ sdkproject/Assets/Mapbox/Unity/Map/Strategies/MapPlacementAtTileCenterStrategy.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 559d97f03cf14bf79a007b09ad4f1336
+timeCreated: 1538658263
