+++ sdkproject/Assets/Mapbox/Unity/Map/Strategies/MapScalingAtUnityScaleStrategy.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6f52911bb8a44f60a4ef4feb3602d386
+timeCreated: 1538658244
