+++ sdkproject/Assets/Mapbox/Unity/Map/Strategies/MapScalingAtWorldScaleStrategy.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: cc8da5d9c56144fc8758276cc931710e
+timeCreated: 1538658236
