+++ sdkproject/Assets/Mapbox/Unity/Map/TileProviders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 59af73b59e6beff4abd67db34649e6e1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
