+++ sdkproject/Assets/Mapbox/Unity/Map/TileProviders/AbstractTileProvider.cs.meta
