+++ sdkproject/Assets/Mapbox/Unity/Map/TileProviders/QuadTreeTileProvider.cs.meta
