+++ sdkproject/Assets/Mapbox/Unity/Map/TileProviders/TileErrorHandler.cs
@@ -1,16 +1,14 @@
-﻿namespace Mapbox.Unity.Map
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Mapbox.Map;
+using UnityEngine;
+using UnityEngine.Events;
+
+namespace Mapbox.Unity.Map.TileProviders
 {
-	using System;
-	using System.Collections.Generic;
-	using System.Linq;
-	using Mapbox.Map;
-	using Mapbox.Unity.Map;
-	using UnityEngine;
-	using UnityEngine.Events;
-
-
 	/// <summary>
-	/// Monobehavior Script to handle TileErrors. 
+	/// Monobehavior Script to handle TileErrors.
 	/// There's an OnTileError event on AbstractMapVisualizer, AbstractTileFactory and UnityTile classes that one can subscribe to to listen to tile errors
 	/// </summary>
 	public class TileErrorHandler : MonoBehaviour
@@ -47,7 +45,7 @@
 				// 1. aborted is always the first exception
 				//    additional exceptions are always caused by the request being aborted
 				//    show all of them as warnings
-				// 2. 'Unable to write data' is another exception associated 
+				// 2. 'Unable to write data' is another exception associated
 				//    with aborted requests: request finshed successfully but
 				//    was aborted during filling of local buffer, also show as warning
 				if (
